+++ sources/TrakDiag/TD_LogWatch8Flags.cpp
@@ -27,12 +27,6 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include <bur/plctypes.h>
-#include <cstring>
-#include <cstdio>
-#include <cctype>
-#include <algorithm>
-
 #ifdef __cplusplus
 	extern "C"
 	{
@@ -41,86 +35,116 @@ SOFTWARE.
 #ifdef __cplusplus
 	};
 #endif
+#include <bur/plctypes.h>
+#include <cstring>
+#include <cstdio>
+
+#include <cctype>
+#include <algorithm>
+
+
+#define STARTUP 0
+#define CHECK_SIGNAL 10
+#define W_DONE 20
+#define ERROR 9999
+
 /* Write status of 8 flags variable into log */
 void TD_LogWatch8Flags(struct TD_LogWatch8Flags* inst)
 {
 	if( inst->Enable ){
-		switch( inst->step ){
-			case 0: /* startup */
-			inst->Busy = 1;
-			std::memset( &inst->fbLogWrite, 0, sizeof(inst->fbLogWrite) );
-			std::strcpy( (char*) inst->fbLogWrite.Name, (char*) inst->LogName );
-			inst->signalOld = !inst->Signal;
-			inst->step = 10;
-			break;
-
-
-			case 10: /* check signal */
-			inst->Valid = 1;
-			if( inst->Signal != inst->signalOld){
-
-				std::strcpy( (char*) inst->fbLogWrite.ObjectID, (char*) inst->SignalName );
-				std::strcpy( (char*) inst->fbLogWrite.Ascii, "" );
-				for( int n = 7; n>=0 ; --n ){
-					if( inst->Flag[n][0] ){  /* flag used ? */
-						char f[sizeof(inst->Flag[0])];
-						bool bit = !!(inst->Signal & 1<<n);
-						bool bitOld = !!(inst->signalOld & 1<<n);
-
-						std::memset( (void*) &f, 0, sizeof(f) );
-						/* use upper case letters for TRUE and lower case for FALSE */
-						if( bit ) {
-							std::transform( &inst->Flag[n][0], &inst->Flag[n][8], f, std::toupper );
-						}
-						else {
-							std::transform( &inst->Flag[n][0], &inst->Flag[n][8], f, std::tolower );
-						}
-						/* add a star '*' as prefix if flag was changed */
-						if( bit != bitOld ){
-							std::strcat( (char*) inst->fbLogWrite.Ascii, "*" );
-						}
-						else {
+		for(;;){
+			switch( inst->step ){
+				case STARTUP: /* startup */
+				inst->Busy = false;
+				inst->Valid = false;
+				inst->Error = false;
+				inst->ErrorID = 0;
+				std::memset( &inst->fbLogWrite, 0, sizeof(inst->fbLogWrite) );
+				std::strcpy( (char*) inst->fbLogWrite.Name, (char*) inst->LogName );
+				TD_LogWrite( &inst->fbLogWrite ); /* reset fb */
+				inst->signalOld = ~inst->Signal; /* force writing when fb is enabled */
+				inst->step = CHECK_SIGNAL;
+				continue; /* save one cycle */
+	
+	
+				case CHECK_SIGNAL: /* check signal */
+				inst->Valid = true;
+				if( inst->Signal != inst->signalOld){
+					std::strcpy( (char*) inst->fbLogWrite.ObjectID, (char*) inst->SignalName );
+					std::strcpy( (char*) inst->fbLogWrite.Ascii, "" );
+					for( int n = 7; n>=0 ; --n ){
+						if( inst->Flag[n][0] ){  /* flag used ? */
+							char f[sizeof(inst->Flag[0])];
+							bool bit = !!(inst->Signal & 1<<n);
+							bool bitOld = !!(inst->signalOld & 1<<n);
+	
+							std::memset( (void*) &f, 0, sizeof(f) );
+							/* use upper case letters for TRUE and lower case for FALSE */
+							if( bit ) {
+								std::transform( &inst->Flag[n][0], &inst->Flag[n][8], f, std::toupper );
+							}
+							else {
+								std::transform( &inst->Flag[n][0], &inst->Flag[n][8], f, std::tolower );
+							}
+							/* add a star '*' as prefix if flag was changed */
+							if( bit != bitOld ){
+								std::strcat( (char*) inst->fbLogWrite.Ascii, "*" );
+							}
+							else {
+								std::strcat( (char*) inst->fbLogWrite.Ascii, " " );
+							}
+							std::strcat( (char*) inst->fbLogWrite.Ascii, f );
 							std::strcat( (char*) inst->fbLogWrite.Ascii, " " );
 						}
-						std::strcat( (char*) inst->fbLogWrite.Ascii, f );
-						std::strcat( (char*) inst->fbLogWrite.Ascii, " " );
 					}
-				}
-				inst->signalOld = inst->Signal;
-				inst->fbLogWrite.EventID = inst->EventID;
-				inst->fbLogWrite.Execute = 1;
-				inst->Busy = 1;
-				TD_LogWrite( &inst->fbLogWrite );
-				if( inst->fbLogWrite.Done ){
-					inst->Busy = 0;
-					inst->fbLogWrite.Execute = 0;
+					inst->signalOld = inst->Signal;
+					inst->fbLogWrite.EventID = inst->EventID;
+					inst->fbLogWrite.Execute = true;
+					inst->Valid = false;
+					inst->Busy = true;
 					TD_LogWrite( &inst->fbLogWrite );
-					inst->step = 10;
+					inst->step = W_DONE;
+				}			
+				break;
+	
+	
+				case W_DONE:
+				if( inst->fbLogWrite.Done ){
+					inst->Busy = false;
+					inst->Error = false;
+					inst->Valid = true;
+					inst->fbLogWrite.Execute = 0; /* reset fb */
+					TD_LogWrite( &inst->fbLogWrite );
+					inst->step = CHECK_SIGNAL;
+					continue; /* save one cycle */
 				}
 				else if( inst->fbLogWrite.Error ){
-					inst->Busy = 0;
-					inst->StatusID = inst->fbLogWrite.StatusID;
-					inst->fbLogWrite.Execute = 0;
+					inst->Busy = false;
+					inst->Valid = false;
+					inst->Error = true;
+					inst->ErrorID = inst->fbLogWrite.ErrorID;
+					inst->fbLogWrite.Execute = 0; /* reset fb */
 					TD_LogWrite( &inst->fbLogWrite );
-					inst->step = 9999;				
+					inst->step = ERROR;				
 				}
-			}			
-			break;
-
-
-			case 9999: /* error */
-			inst->Valid = 0;
-			inst->Busy = 0;
-			inst->Error = 1;
-			break;
-
-		}
+				else {
+					TD_LogWrite( &inst->fbLogWrite );
+				}
+				break;
+				
+	
+				case ERROR: /* error */
+				break;
+	
+			}
+			break; /* leave endless loop */
+		} /* end for(;;) */
 	}
 	else {
-		inst->Busy = 0;
-		inst->Valid = 0;
-		inst->StatusID = 0;
-		inst->Error = 0;
-		inst->step = 0;
+		inst->Busy = false;
+		inst->Valid = false;
+		inst->ErrorID = 0;
+		inst->Error = false;
+		inst->step = STARTUP;
 	}
 }